Added an id to every component when saving for reconstruction purposes when loading a saved file.
Added a way to retrieve the output of a connector.

Assign an id to every component when saving to a json file. Save the state of sources, whether they are on or off in the json files so reloaded session will have the states restored.

Save the connector state in a json file including the id of the input, output and which of the input connect points.

Build the connectors on load reconstruct previous sessions connectors.